[JOLT-166] RV64

+++ jolt-core/src/zkvm/lookup_table/suffixes/left_is_zero.rs
@@ -6,7 +6,7 @@ use super::SparseDenseSuffix;
 pub enum LeftOperandIsZeroSuffix {}
 
 impl SparseDenseSuffix for LeftOperandIsZeroSuffix {
-    fn suffix_mle(b: LookupBits) -> u32 {
+    fn suffix_mle(b: LookupBits) -> u64 {
         let (x, _) = b.uninterleave();
         (u64::from(x) == 0).into()
     }
