+++ jolt-core/src/zkvm/lookup_table/suffixes/right_shift_helper.rs
@@ -8,7 +8,7 @@ use super::SparseDenseSuffix;
 pub enum RightShiftHelperSuffix {}
 
 impl SparseDenseSuffix for RightShiftHelperSuffix {
-    fn suffix_mle(b: LookupBits) -> u32 {
+    fn suffix_mle(b: LookupBits) -> u64 {
         let (_, y) = b.uninterleave();
         1 << y.leading_ones()
     }
